+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Fix detection of another instance already running in `join-pool` command.
+- Fix farmer not exiting properly on ctrl-c.
 
 ## [1.9.0] - 2023-10-13